threadcommon address->char | reading fixed		   | writing fixed		   | send request modified	   | casted mov_average <-------

+++ src/measuring.c
@@ -44,9 +44,9 @@ void takeoutFromQueue(void *arg)
                     TAILQ_REMOVE(&common->head,tempPacket,entries);
                     pthread_mutex_unlock(&common->mutex);
 
-                    devices[tempPacket->address].k_next=*(tempPacket->data);
+                    devices[(int)tempPacket->address].k_next=*(tempPacket->data);
 
-                    temp=mov_average(&devices[tempPacket->address],common->members);
+                    temp=mov_average(&devices[(int)tempPacket->address],common->members);
                     finalResult=moving_hysteresis(common->Delta,temp);
 
                     syslog(LOG_INFO,"Measured temperature from %d address of device with moving average and moving hysteresis :%.2f\n",
